+++ src/c/utests/data/data.c
@@ -2763,6 +2763,23 @@ static void test_data_map_get_list (void)
   iot_data_free (map);
 }
 
+static void test_data_map_take (void)
+{
+  iot_data_t * list_key = iot_data_alloc_string ("list", IOT_DATA_REF);
+  iot_data_t * string_key = iot_data_alloc_string ("str", IOT_DATA_REF);
+  iot_data_t * list = iot_data_alloc_list ();
+  iot_data_t * map = iot_data_alloc_map (IOT_DATA_STRING);
+  iot_data_map_add (map, string_key, iot_data_alloc_string ("string", IOT_DATA_REF));
+  iot_data_map_add (map, list_key, list);
+  iot_data_t * found = iot_data_map_take (map, list_key);
+  CU_ASSERT (iot_data_map_size (map) == 1u)
+  CU_ASSERT (found == list)
+  const iot_data_t * found2 = iot_data_map_get (map, list_key);
+  CU_ASSERT (found2 == NULL)
+  iot_data_free (found);
+  iot_data_free (map);
+}
+
 static void test_data_map_get_array (void)
 {
   iot_data_t * array_key = iot_data_alloc_string ("array", IOT_DATA_REF);
@@ -5728,6 +5745,8 @@ void cunit_data_test_init (void)
   CU_add_test (suite, "data_map_iter_replace", test_data_map_iter_replace);
   CU_add_test (suite, "data_map_remove", test_data_map_remove);
   CU_add_test (suite, "data_map_get_list", test_data_map_get_list);
+  CU_add_test (suite, "data_map_take", test_data_map_take);
+
   CU_add_test (suite, "data_map_add_unused", test_data_map_add_unused);
   CU_add_test (suite, "data_map_get_array", test_data_map_get_array);
   CU_add_test (suite, "data_string_vector", test_data_string_vector);