+++ RELEASE_NOTES.md
@@ -24,3 +24,4 @@
 - Added additional stats collection and publishing for the scheduler component
 - Data memory management update for better performance
 - Added functions `iot_data_list_head_push_list` and `iot_data_list_tail_push_list` to support list merging
+- Added function `iot_data_map_take` to take a value from a map